Ajuste ProdutoDTO (Preco de decimal para double); Teste de unidade PUT em ProdutosController.

+++ APICatalogo/Pagination/ProdutosFIltroPreco.cs
@@ -2,7 +2,7 @@
 {
     public class ProdutosFIltroPreco : QueryStringParameters
     {
-        public decimal? Preco { get; set; }
+        public double? Preco { get; set; }
         public string? PrecoCriterio { get; set; }
     }
 }
